Move interop/wasi to tests/wasi
